requirements.txt
python-dotenv==1.0.0
selenium==4.16.0
webdriver-manager==4.0.1
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
requests==2.31.0
google-auth-oauthlib==0.8.0
